Removed structure mess, added SetupActivity
