+++ app/src/main/java/com/example/traccite/SetupActivity.java
@@ -1,0 +1,25 @@
+package com.example.traccite;
+
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.content.Context;
+import android.content.Intent;
+import android.os.Bundle;
+
+public class SetupActivity extends AppCompatActivity {
+
+  // Logcat: Logging Tag
+  private static final String TAG = "SetupActivity";
+
+  @NonNull
+  public static Intent createIntent(@NonNull Context context) {
+    return new Intent(context, SetupActivity.class);
+  }
+
+  @Override
+  protected void onCreate(Bundle savedInstanceState) {
+    super.onCreate(savedInstanceState);
+    setContentView(R.layout.activity_setup);
+  }
+}
