--- app/src/main/res/layout/activity_auth.xml
@@ -1,117 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-  xmlns:app="http://schemas.android.com/apk/res-auto"
-  xmlns:tools="http://schemas.android.com/tools"
-  android:layout_width="match_parent"
-  android:layout_height="match_parent"
-  tools:context=".AuthActivity">
-
-  <TextView
-    android:id="@+id/title"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    android:layout_marginTop="32dp"
-    android:text="Before Continuing..."
-    android:textAppearance="@style/TextAppearance.AppCompat.Large"
-    android:textStyle="bold"
-    app:layout_constraintBottom_toBottomOf="parent"
-    app:layout_constraintEnd_toEndOf="parent"
-    app:layout_constraintHorizontal_bias="0.075"
-    app:layout_constraintStart_toStartOf="parent"
-    app:layout_constraintTop_toTopOf="parent"
-    app:layout_constraintVertical_bias="0.022" />
-
-  <TextView
-    android:id="@+id/subtitle"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    android:text="Tell us more about you"
-    android:textAppearance="@style/TextAppearance.AppCompat.Medium"
-    app:layout_constraintBottom_toBottomOf="parent"
-    app:layout_constraintEnd_toEndOf="parent"
-    app:layout_constraintHorizontal_bias="0.069"
-    app:layout_constraintStart_toStartOf="parent"
-    app:layout_constraintTop_toBottomOf="@+id/title"
-    app:layout_constraintVertical_bias="0.0" />
-
-  <EditText
-    android:id="@+id/nric_number"
-    android:focusable="true"
-    android:focusableInTouchMode="true"
-    android:layout_width="380dp"
-    android:layout_height="wrap_content"
-    android:layout_marginTop="32dp"
-    android:layout_marginBottom="584dp"
-    android:ems="10"
-    android:fontFamily="monospace"
-    android:hint="NRIC"
-    android:inputType="textPersonName"
-    android:textColor="#000"
-    app:layout_constraintBottom_toBottomOf="parent"
-    app:layout_constraintEnd_toEndOf="parent"
-    app:layout_constraintHorizontal_bias="0.516"
-    app:layout_constraintStart_toStartOf="parent"
-    app:layout_constraintTop_toBottomOf="@+id/subtitle"
-    app:layout_constraintVertical_bias="0.0" />
-
-  <EditText
-    android:id="@+id/full_name"
-    android:layout_width="380dp"
-    android:layout_height="wrap_content"
-    android:layout_marginTop="8dp"
-    android:ems="10"
-    android:hint="Full Name"
-    android:inputType="textPersonName"
-    android:textColor="#000"
-    app:layout_constraintBottom_toBottomOf="parent"
-    app:layout_constraintEnd_toEndOf="parent"
-    app:layout_constraintHorizontal_bias="0.516"
-    app:layout_constraintStart_toStartOf="parent"
-    app:layout_constraintTop_toBottomOf="@+id/nric_number"
-    app:layout_constraintVertical_bias="0.0" />
-
-  <EditText
-    android:id="@+id/contact_number"
-    android:layout_width="380dp"
-    android:layout_height="wrap_content"
-    android:layout_marginTop="8dp"
-    android:ems="10"
-    android:enabled="false"
-    android:hint="Contact Number"
-    android:inputType="phone"
-    android:textColor="#000"
-    app:layout_constraintBottom_toBottomOf="parent"
-    app:layout_constraintEnd_toEndOf="parent"
-    app:layout_constraintHorizontal_bias="0.516"
-    app:layout_constraintStart_toStartOf="parent"
-    app:layout_constraintTop_toBottomOf="@+id/full_name"
-    app:layout_constraintVertical_bias="0.0" />
-
-  <Switch
-    android:id="@+id/resident_of_singapore"
-    android:layout_width="380dp"
-    android:layout_height="wrap_content"
-    android:layout_marginTop="8dp"
-    android:checked="true"
-    android:text="I am a resident of Singapore"
-    app:layout_constraintBottom_toBottomOf="parent"
-    app:layout_constraintEnd_toEndOf="parent"
-    app:layout_constraintHorizontal_bias="0.516"
-    app:layout_constraintStart_toStartOf="parent"
-    app:layout_constraintTop_toBottomOf="@+id/contact_number"
-    app:layout_constraintVertical_bias="0.0" />
-
-  <Button
-    android:id="@+id/continue_button"
-    android:layout_width="380dp"
-    android:layout_height="wrap_content"
-    android:layout_marginTop="8dp"
-    android:text="Continue"
-    app:layout_constraintBottom_toBottomOf="parent"
-    app:layout_constraintEnd_toEndOf="parent"
-    app:layout_constraintHorizontal_bias="0.483"
-    app:layout_constraintStart_toStartOf="parent"
-    app:layout_constraintTop_toBottomOf="@+id/resident_of_singapore"
-    app:layout_constraintVertical_bias="0.0" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
