+++ app/src/main/res/layout/activity_setup.xml
@@ -1,0 +1,120 @@
+<?xml version="1.0" encoding="utf-8"?>
+<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
+  xmlns:tools="http://schemas.android.com/tools"
+  android:layout_width="match_parent"
+  android:layout_height="match_parent"
+  android:background="@color/gray200"
+  android:orientation="vertical"
+  android:scrollbarThumbVertical="@null"
+  android:scrollbarThumbHorizontal="@null"
+  android:overScrollMode="never"
+  tools:context=".SetupActivity">
+
+  <LinearLayout
+    android:padding="10dp"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical">
+
+    <LinearLayout
+      android:layout_width="match_parent"
+      android:layout_height="wrap_content"
+      android:orientation="horizontal"
+      android:paddingLeft="10dp"
+      android:paddingTop="10dp"
+      android:paddingRight="10dp">
+
+      <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:fontFamily="@font/montserrat_bold"
+        android:text="Before we continue..."
+        android:textColor="@color/gray700"
+        android:textSize="25dp"
+        android:textStyle="bold" />
+    </LinearLayout>
+
+    <LinearLayout
+      android:layout_width="match_parent"
+      android:layout_height="wrap_content"
+      android:orientation="horizontal"
+      android:paddingLeft="10dp"
+      android:paddingRight="10dp">
+
+      <TextView
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:fontFamily="@font/montserrat_medium"
+        android:text="Tell us more about you!"
+        android:textColor="@color/gray500"
+        android:textSize="15dp"
+        android:textStyle="normal" />
+    </LinearLayout>
+
+    <LinearLayout
+      android:layout_width="match_parent"
+      android:layout_height="wrap_content"
+      android:layout_marginTop="30dp"
+      android:orientation="vertical"
+      android:paddingLeft="10dp"
+      android:paddingRight="10dp">
+
+      <EditText
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginBottom="15dp"
+        android:background="@color/gray300"
+        android:hint="NRIC Number"
+        android:inputType="textNoSuggestions"
+        android:maxLength="9"
+        android:padding="10dp"
+        android:textColor="@color/black" />
+
+      <EditText
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginBottom="15dp"
+        android:background="@color/gray300"
+        android:hint="Full Name"
+        android:inputType="textNoSuggestions"
+        android:padding="10dp"
+        android:textColor="@color/black" />
+
+      <EditText
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginBottom="15dp"
+        android:background="@color/gray300"
+        android:hint="Contact Number"
+        android:inputType="number"
+        android:maxLength="10"
+        android:padding="10dp"
+        android:textColor="@color/black" />
+
+      <Switch
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:checked="true"
+        android:fontFamily="@font/montserrat_light"
+        android:paddingTop="10dp"
+        android:paddingBottom="10dp"
+        android:text="I am a resident of Singapore" />
+    </LinearLayout>
+
+    <LinearLayout
+      android:layout_width="match_parent"
+      android:layout_height="wrap_content"
+      android:layout_marginTop="15dp"
+      android:orientation="horizontal"
+      android:paddingLeft="10dp"
+      android:paddingRight="10dp">
+
+      <Button
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:background="@color/black"
+        android:text="Continue"
+        android:textColor="@color/white" />
+    </LinearLayout>
+  </LinearLayout>
+</ScrollView>
