+++ app/src/main/res/values/styles.xml
@@ -7,6 +7,7 @@
     <item name="colorAccent">@color/colorAccent</item>
   </style>
 
+  <!-- Remove when we no longer need FirebaseUI -->
   <style name="AuthTheme" parent="FirebaseUI">
     <item name="windowActionBar">false</item>
     <item name="windowNoTitle">true</item>
